fixes buttons not showing up
Script broke due to youtube changing its page layout. Unfortunately
this script WILL break again, as soon as Google fiddles with parts of
the page the script relies on.

Nature of the problem user scripts were created for, building a house on
a patch of mud / making assumptions about things we don't control.
A user script will only work solidly at the instant of a commit.